Project nearly entering beta stage.

+++ lib/compat/src/attiny.c
@@ -1,7 +1,8 @@
 #include <avr/io.h>
+#include <avr/eeprom.h>
 
 // Turn this into interrupt later
-int adc(uint8_t pin)
+uint16_t adc(uint8_t pin)
 {
   ADCSRA |= (1 << ADSC); // Start ADC conversion. Turns off in hardware after completion.
   ADMUX = pin;
